+++ custom_components/cubecoders/coordinator.py
@@ -8,7 +8,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.exceptions import ConfigEntryAuthFailed
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
-from .api import AmpApiClientAuthenticationError, AmpApiClientError, AmpExtendedInstance
+from .api import AmpApiClientAuthenticationError, AmpApiClientError, AmpCoordinatorData
 
 if TYPE_CHECKING:
     from .data import AmpData
@@ -17,7 +17,7 @@ type AMPConfigEntry = ConfigEntry[AmpData]
 
 
 # https://developers.home-assistant.io/docs/integration_fetching_data#coordinated-single-api-poll-for-data-for-all-entities
-class AmpDataUpdateCoordinator(DataUpdateCoordinator[dict[int, AmpExtendedInstance]]):
+class AmpDataUpdateCoordinator(DataUpdateCoordinator[AmpCoordinatorData]):
     """Class to manage fetching data from the API."""
 
     config_entry: AMPConfigEntry